hero component updated
